requirements.txt
Django
django-ckeditor
django-js-asset
Pillow
sqlparse
sphinx
sphinx_rtd_theme
typing_extensions
tzdata
qrcode
